Delete EthereumMessageSigner
AppChain should not handle Ethereum signing.
